+++ pkg/pub_worker/test/dockerized_end2end_test.dart
@@ -76,6 +76,30 @@ void main() {
               ? '[no log.txt]'
               : utf8.decode(gzip.decode(logTxtBytes));
 
+          // verify unexpected failures
+          final unexpectedFragments = [
+            'Failed to run',
+            'certificate verification failed',
+            'access denied',
+          ];
+          for (final fragment in unexpectedFragments) {
+            if (logTxt.toLowerCase().contains(fragment.toLowerCase())) {
+              fail('Log contains "$fragment":\n$logTxt');
+            }
+          }
+
+          // verify log for sandboxed executions
+          final expectedFragments = [
+            'build/sandbox_runner /home/worker/dart/stable/bin/dart pub',
+            'build/sandbox_runner git',
+            'sandbox_runner /home/worker/dartdoc/build/dartdoc',
+            if (packagesWithScreenshots.contains(package))
+              'build/sandbox_runner webpinfo',
+          ];
+          for (final fragment in expectedFragments) {
+            expect(logTxt, contains(fragment));
+          }
+
           final docIndex = result.index.lookup('doc/index.html');
           expect(
             docIndex,
@@ -91,6 +115,16 @@ void main() {
           );
           final report = summary.report!;
           expect(report.maxPoints, greaterThan(100));
+
+          // check the presence of a license
+          expect(summary.tags, isNotNull);
+          expect(
+            summary.tags!.any(
+              (tag) => tag.startsWith('license:') && tag != 'license:unknown',
+            ),
+            isTrue,
+            reason: summary.tags.toString(),
+          );
 
           if (packagesWithScreenshots.contains(package)) {
             expect(